增加自动提交变更单，自动比对old.zip 等功能

+++ routes/svn.js
@@ -23,7 +23,7 @@ var getCommitPath  = function(){
     var now   = new Date();
     var year = now.getFullYear().toString() ;
     var month = (now.getMonth() + 1).toString();
-    return  year +"-" + month+'11/';
+    return  year +"-" + month+'/';
 }
 var getFilePath = function(path){
     if(!path){
@@ -81,14 +81,16 @@ var copyFile = function(sourceDir, destDir, fileName){
     //})
     console.log("复制完成");
 }
-router.post("/commitChangeRar/:attachmentId",function(req, res){
+router.post("/commitChangeRar/",function(req, res){
     Tool.getCookieUser(req, res);
-    var attachmentId = req.params.attachmentId;
+    //var attachmentId = req.params.attachmentId;
+    var attachmentId = req.body.attachmentId;
     AttaSql.searchAttaAndSvn(attachmentId,changeRarSvnId,function(msg,result){
         var path = "/attachment/changeRar/"+ getCommitPath();
         if((msg == "err")||(result == null)){
-            req.session.error = "查找附件信息时时发生错误【"+result+"】，请记录并联系管理员";
-            return res.redirect("/admin/findAttaHistory/");
+            return ajaxReturn(req,res,"err","查找附件信息时时发生错误，请联系管理员");
+            //req.session.error = "查找附件信息时时发生错误【"+result+"】，请记录并联系管理员";
+            //return res.redirect("/admin/findAttaHistory/");
         }
         //console.log("result.fileName", process.cwd());
         var attaPath = getFilePath(result.fileUri);
@@ -96,6 +98,11 @@ router.post("/commitChangeRar/:attachmentId",function(req, res){
         var newName = result.fileName;
         if(!fs.existsSync(curPath+attaPath.path+attaPath.file)){
             return ajaxReturn(req,res,"err","附件不存在，请联系管理员");
+            //req.session.error = "附件提交至svn上出错:["+attaPath.path+attaPath.file+"]不存在，请记录并联系管理员";
+            //return res.redirect("/admin/findAttaHistory/");
+        }
+        if(!fs.existsSync(curPath+path+attaPath.file)&&!fs.existsSync(curPath+path+newName)){
+            return ajaxReturn(req,res,"err","上传副本不存在，请先创建");
             //req.session.error = "附件提交至svn上出错:["+attaPath.path+attaPath.file+"]不存在，请记录并联系管理员";
             //return res.redirect("/admin/findAttaHistory/");
         }
@@ -110,19 +117,29 @@ router.post("/commitChangeRar/:attachmentId",function(req, res){
                 var option = result_svn;
                 svn = new Svn(option);
                 console.log("commit success");
-                svn.commitChangeRar(curPath+path, curPath + attaPath.path,  attaPath.file, newName, result.svnUri, message, function (msg, result) {
+                svn.commitChangeRar(curPath+path, curPath + attaPath.path,  attaPath.file, newName, result.svnUri, message, function (msg, com_result) {
                     if (msg == "err") {
-                        return ajaxReturn(req,res,"err","附件提交至svn上出错【" + result + "】，请记录并联系管理员");
-                        //req.session.error = "附件提交至svn上出错【" + result + "】，请记录并联系管理员";
-                        //return res.redirect("/admin/findAttaHistory/");
+                        return ajaxReturn(req,res,"err","附件提交至svn上出错【" + com_result + "】，请记录并联系管理员");
+
                     }
-                    return ajaxReturn(req,res,"success","附件提交至svn上成功");
+                    AttaSql.commitRar(attachmentId,function(msg,cr_commit){
+                         if(msg == "success"){
+                             console.log("附件提交至svn上成功");
+                             return ajaxReturn(req,res,"success","附件提交至svn上成功");
+                         }
+                        else{
+                              console.log("附件提交成功，数据库记录出错！请联系管理员");
+                              return ajaxReturn(req,res,"err","附件提交成功，数据库记录出错！请联系管理员");
+                         }
+                    });
+                    //return ajaxReturn(req,res,"success","附件提交至svn上成功");
                     //return res.redirect("/admin/findAttaHistory/");
                 });
             }
         });
     });
 });
+/*将变更单复制到指定路径下*/
 router.post("/copyRar/",function(req, res) {
     Tool.getCookieUser(req, res);
     var attachmentId = req.body.attachmentId;
@@ -131,25 +148,26 @@ router.post("/copyRar/",function(req, res) {
     AttaSql.searchAttaAndSvn(attachmentId, changeRarSvnId, function (msg, result) {
         var path = "/attachment/changeRar/" + getCommitPath();
         if ((msg == "err") || (result == null)) {
-            req.session.error = "查找附件信息时时发生错误【" + result + "】，请记录并联系管理员";
+            console.error("查找附件信息时时发生错误，请记录并联系管理员");
+            return  ajaxReturn(req,res,"err","查找附件信息时时发生错误，请记录并联系管理员");
+            //req.session.error = "查找附件信息时时发生错误【" + result + "】，请记录并联系管理员";
             //return res.redirect("/admin/findAttaHistory/");
         }
         //console.log("result.fileName", process.cwd());
         var attaPath = getFilePath(result.fileUri);
         var message = getFileName(result.fileName);//文件名作为svn上传的message
         var newName = result.fileName;
-        if (!fs.existsSync(curPath + attaPath.path + attaPath.file)) {
+        if (!fs.existsSync(curPath + attaPath.path + attaPath.file)) {//变更单是否存在
             //req.session.error = "附件复制出错:[" + attaPath.path + attaPath.file + "]不存在，请记录并联系管理员";
             //return res.redirect("/admin/findAttaHistory/");
-            var queryObj = url.parse(req.url,true).query;
-            var jsonStr;
-            jsonStr = '{"sucFlag":"err","message":"拷贝失败"}';
-            return  res.send(queryObj.callback+'(\'' + jsonStr + '\')');
+            console.error("变更单不存在");
+            return  ajaxReturn(req,res,"err","拷贝失败");
         }
         dao.getSvnUser(function (s_msg, result_svn) {
             if (s_msg === "err") {
                 console.log("getSvnUser：" + err);
-                req.session.error = "查找附件信息时时发生错误【" + result + "】，请记录并联系管理员";
+                return  ajaxReturn(req,res,"err","查找svn账号信息时发生错误，请记录并联系管理员");
+                //req.session.error = "查找附件信息时时发生错误【" + result + "】，请记录并联系管理员";
                 //return res.redirect("/admin/findAttaHistory/");
             }
             else if (s_msg = "success") {
@@ -162,29 +180,46 @@ router.post("/copyRar/",function(req, res) {
                 var newName = result.fileName;
                 svn.svnExists(filePath, versionUri, "变更单", function (e_msg, data) {
                     if (e_msg == "err") {
+                        console.error("创建文件夹失败svnExists:" + data);
                         //req.session.error = "附件拷贝出错，请记录并联系管理员";
-                        var queryObj = url.parse(req.url,true).query;
-                        var jsonStr;
-                        jsonStr = '{"sucFlag":"err","message":"拷贝失败"}';
-                        return  res.send(queryObj.callback+'(\'' + jsonStr + '\')');
+                        return  ajaxReturn(req,res,"err","拷贝失败");
                         //return res.redirect("/admin/findAttaHistory/");
                     }
                     if (!fs.existsSync(oldPath + fileName)) {
-                        //console.log(oldPath + fileName, "不存在！！");
+                        console.error(oldPath + fileName, "不存在！！");
                         //return;
                         return ajaxReturn(req,res,"err","附件拷贝【" + result + "】不存在，请记录并联系管理员");
                     }
                     console.log("copy:", filePath + fileName, "  ", filePath + newName);
+                    if(fs.existsSync(filePath + newName)||fs.existsSync(filePath + fileName )){
+                        return  ajaxReturn(req,res,"success","已复制，无需重复复制");
+                    };
                     copyFile(oldPath, filePath, fileName);
-                    var queryObj = url.parse(req.url,true).query;
-                    var jsonStr;
-                    jsonStr = '{"sucFlag":"success","message":"拷贝成功"}';
-                    res.send(queryObj.callback+'(\'' + jsonStr + '\')');
+                    return  ajaxReturn(req,res,"success","复制成功");
                     //fs.renameSync(filePath + fileName, filePath + newName);
                 });
             }
         });
     });
 });
-
+router.get("/commitRarManual/:attachmentId",function(req, res){
+    Tool.getCookieUser(req, res);
+    //var attachmentId = req.params.attachmentId;
+   if(!req.cookies.user.isAdmin){
+       res.redirect("/");
+   } ;
+    var attachmentId = req.params.attachmentId;
+    AttaSql.commitRar(attachmentId,function(msg,cr_commit){
+        if(msg == "success"){
+            console.log("附件提交至svn上成功");
+            return  res.redirect("/admin/findAttaHistory");;
+            //return ajaxReturn(req,res,"success","附件提交至svn上成功");
+        }
+        else{
+            console.log("附件提交成功，数据库记录出错！请联系管理员");
+            req.session.error = "";
+           res.redirect("/admin/findAttaHistory");
+        }
+    });
+});
 module.exports  = router;
